requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Deep learning and computer vision
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0
diffusers>=0.21.0
transformers>=4.30.0

# Image processing
Pillow>=9.0.0

# OpenAI API
openai>=1.0.0

# Plotting and visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0

# File handling
pathlib2>=2.3.0

# Optional: For faster attention mechanisms
# xformers>=0.0.20

# Development and testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
